# Improve performance of LazyBridgeOptimizer